requirements/development.txt
asgiref==3.8.1
attrs==24.3.0
autopep8==2.3.1
Django==5.1.4
django-autoslug==1.9.9
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-dirtyfields==1.9.5
django-filter==24.3
django-silk==5.3.2
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
gprof2dot==2024.6.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
pillow==11.0.0
pycodestyle==2.12.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
